// WebSystem/WebSystem/ClientApp/src/models/Formas/COB/dp05a110.ts
export interface IDp05a110 {
        id?: number;
        codigo: string;
        cod_uees?: string;
	      pc_ncta?: string;
        empcli: string;
        nomcli?: string;
        tipopre?: string;
        ruc?: string;
        direccion?: string;
        direccion2?: string;
        fono1?: string;
        fono2?: string;
        vendedor?: string;
        diasp?: number;
        fecha_ultv?: Date;
        valor_ultv?: number;
        balance?: number;
        abierto?: number;
        estado?: string;
        pordes?: number;
        fecha_ing?: Date;
        comenta?: string;
        fecha_ultc?: Date;
        valor_ultc?: number;
        zona?: string;
        sector?: string;
        credito?: number;
        tipo_cli?: string;
        ciudad?: string;
        fecha_nac?: Date;
        sexo?: string;
        titulo?: string;
        colegio?: string;
        provin_col?: string;
        ciudad_col?: string;
        parentezco?: string;
        nombre_rep?: string;
        lugarnacre?: string;
        ocupa_rep?: string;
        cargo_rep?: string;
        empre_rep?: string;
        dire_rep?: string;
        fonos_rep?: string;
        email_rep?: string;
        facultad?: string;
        programa?: string;
        especiali?: string;
        clasifica?: string;
        clicodint?: string;
        patrocina?: string;
        patcodint?: string;
        tipotrans?: string;
        cl_precio?: number;
        catego?: number;
        email?: string;
        es_contrib?: string;
        estado_tar?: string;
        digito_tar?: string;
        profesion?: string;
        recipromo?: string;
        usoprod?: string;
        prodpara?: string;
        tpusoprod?: string;
        nhijos?: number;
        st_civil?: string;
        tipocli?: string;
        tpdoc?: string;
        cuota_1?: number;
        tipo_1?: string;
        cuota_2?: number;
        tipo_2?: string;
        codmain?: string;
        replegal?: string;
        repced?: string;
        repfono?: string;
        fax?: string;
        webpage?: string;
        paisnom?: string;
        provincia?: string;
        planea?: string;
        marca?: string;
        marca2?: string;
        origen?: string;
        seriefa?: string;
        autofa?: string;
        f_emifa?: Date;
        f_cadufa?: Date;
        conse?: string;
        tipo_id?: string;
        es_estatal?: string;
        canton?: string;
        parroquia?: string;
        estcivil?: string;
        origening?: string;
        es_parterel?: number;
        retiene?: string;
        antic_ncta?: string;
        ciudad_respal?: string;
        esparterel?: string;
        tp_cliente?: string;
        t_cobro?: string;
        user_crea?: string;
        fec_crea?: Date;
        user_mod?: string;
        fec_mod?: Date;
}